Move NegotiateTrade class to mission task package
Ref #1550
