+++ mars-sim-core/src/main/java/com/mars_sim/core/person/ai/mission/Trade.java
@@ -21,9 +21,9 @@ import com.mars_sim.core.goods.Good;
 import com.mars_sim.core.logging.SimLogger;
 import com.mars_sim.core.map.location.LocalPosition;
 import com.mars_sim.core.mission.objectives.TradeObjective;
+import com.mars_sim.core.mission.task.NegotiateTrade;
 import com.mars_sim.core.person.Person;
 import com.mars_sim.core.person.ai.SkillType;
-import com.mars_sim.core.person.ai.task.NegotiateTrade;
 import com.mars_sim.core.person.ai.task.Walk;
 import com.mars_sim.core.person.ai.task.WalkingSteps;
 import com.mars_sim.core.person.ai.task.util.Worker;
